# mwavuli-backend/infra/openapi.yaml
openapi: 3.1.0
info:
  title: mwavuli API
  version: "0.1.0"
  description: >
    Tree identification, mapping, and community backend. Public endpoints expose
    only fuzzy locations; exact coordinates require ownership/staff and are
    audited.
servers:
  - url: https://api.mwavuli.app
  - url: http://localhost:8080
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
...
    Error:
      type: object
      properties: { error: { type: string }, message: { type: string } }
paths:
...
  /v1/auth/register:
    post:
      summary: Create an account (13+ / COPPA gate)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, username, password, displayName, birthYear, acceptTos]
              properties:
                email: { type: string, format: email }
                username: { type: string }
                password: { type: string, minLength: 8 }
                displayName: { type: string }
                birthYear: { type: integer }
                acceptTos: { type: boolean, enum: [true] }
      responses:
        '200': { description: user + tokens }
        '400': { description: under 13 / invalid, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: email/username taken }
...
  /v1/me/export:
    post:
      summary: GDPR data export (JSON/CSV)
      parameters: [ { name: format, in: query, schema: { type: string, enum: [json, csv], default: json } } ]
      responses: { '200': { description: export archive } }
...